refactor: 함수 및 파일 구조 변경
